requirements.txt
nltk==3.8.1
numpy==1.21.5
PyYAML==6.0
torch==1.12.1
tqdm==4.64.0
